requirements/local.txt
# The requirement file to be used on local development system
-r base.txt
coverage==3.7.1
django-debug-toolbar==1.2
